+++ src/helpers/Assets.php
@@ -183,7 +183,13 @@ class Assets
      */
     public static function filename2Title(string $filename): string
     {
-        return StringHelper::upperCaseFirst(implode(' ', StringHelper::toWords($filename, false, true)));
+        $title = StringHelper::upperCaseFirst(implode(' ', StringHelper::toWords($filename, false, true)));
+
+        if (strlen($title) > 255) {
+            $title = rtrim(substr($title, 255), ' ');
+        }
+
+        return $title;
     }
 
     /**